@@ -1,18 +1,29 @@
-import React from 'react'
+import React from "react";
+import styled from "styled-components";
 
-function TodoItem(props) {
-    const completedStyle = {
-        color: 'grey',
-        textDecoration: 'line-through'
-    }
+const Item = styled.li`
+  color: ${({ theme }) => theme.colors.grayscale.contrast};
+  background-color: ${({ theme }) => theme.colors.grayscale.light};
+  padding: 5px 0;
+  line-height: 1.6;
+  text-decoration: ${(props) => (props.$complete ? "line-through" : "none")};
 
-    return (
-        <li 
-            style={props.item.completed ? completedStyle: null}
-            onClick={() => props.handleClick(props.item.id)}>
-            {props.item.text}
-        </li>
-    )
+  &:nth-child(2n) {
+    color: ${({ theme }) => theme.colors.grayscale.main};
+    background-color: ${({ theme }) => theme.colors.grayscale.contrast};
+  }
+`;
+
+function TodoItem({ item, handleClick }) {
+  return (
+    <Item
+      id={item.id}
+      $complete={item.completed}
+      onClick={(e) => handleClick(e)}
+    >
+      {item.text}
+    </Item>
+  );
 }
 
-export default TodoItem
+export default TodoItem;
